Fix computation of the suffix for executable files.
Executable files were being given the suffix ".exe"
even on non-windows platforms. Ordinarily this is not
a problem, but with Windows Subsystem for Linux
this sometimes causes Windows to interpret the file
as a Windows executable rather than a Linux executable.
One one case (while using opam to install opam module core_kernel)
this was causing a test C program to get a kill signal
from the system.

Signed-off-by: Richard L Ford <richardlford@gmail.com>